tmux: Add binary finding
Now searches $PATH to find which tmux to execute rather than hardcoding
it. This is better to work across systems where they are installed in
different locations. Obviously, it skips its self in the search result
to prevent the infinite loop.

This also changes the script from being sh-compatible to requiring bash